fix(eslint): Prepare eslint for rollup 3 and upgrade eslint to newest version (#1309)
Prior to these changes, the eslint plugin was using an older
version of eslint that relied on a class called CLIEngine.
In newer versions this has now been deprecated and so upgrading
versions required also updating to using ESLint internally
instead. Ultimately this doesn't change the usage of the plugin
but some of the option names have changed for ESLint (notably
configFile needs to change to overrideConfigFile).

The main driving factor for this change is to prevent conflicts
that happen when using @rollup/plugin-typescript and this plugin
together. Without this update, eslint would run on the transpiled
code and so the error messages weren't as helpful. This update
will allow the two plugins to work well together and provide
helpful information about errors and warnings.

This change also includes some necessary package updates to prepare
eslint for Rollup 3.

BREAKING CHANGES:
- options structure has changed due to moving from CLIEngine t
  ESlint. Review the new supported options below:
https://eslint.org/docs/latest/developer-guide/nodejs-api#-new-eslintoptions
- requires Node 14